requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
Django==5.1.2
django-filter==24.3
django-rest==0.8.7
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
macaroonbakery==1.3.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
openai==1.37.0
protobuf==3.12.4
psycopg2==2.9.10
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydbus==0.6.0
pygame==2.5.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-decouple==3.8
python-dotenv==1.0.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.5
wadllib==1.3.6
xdg==5.0.0
zipp==1.0.0
